+++ backend/src/test/java/com/staccato/moment/service/MomentServiceTest.java
@@ -17,6 +17,8 @@ import org.springframework.mock.web.MockMultipartFile;
 import com.staccato.ServiceSliceTest;
 import com.staccato.exception.ForbiddenException;
 import com.staccato.exception.StaccatoException;
+import com.staccato.fixture.moment.CommentFixture;
+import com.staccato.fixture.moment.MomentFixture;
 import com.staccato.member.domain.Member;
 import com.staccato.member.repository.MemberRepository;
 import com.staccato.memory.domain.Memory;
@@ -24,8 +26,6 @@ import com.staccato.memory.repository.MemoryRepository;
 import com.staccato.moment.domain.Comment;
 import com.staccato.moment.domain.Moment;
 import com.staccato.moment.domain.MomentImages;
-import com.staccato.moment.fixture.CommentFixture;
-import com.staccato.moment.fixture.MomentFixture;
 import com.staccato.moment.repository.CommentRepository;
 import com.staccato.moment.repository.MomentImageRepository;
 import com.staccato.moment.repository.MomentRepository;
